requirements.txt
asgiref==3.3.1
Django==3.0.10
djangorestframework==3.12.4
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
